+++ demo/import-html.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>html-loader</title>
+</head>
+<body>
+    <script src="../dist/berial.js"></script>
+    <script>
+        
+    </script>
+</body>
+</html>
+
+
+<!-- <!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>import-html</title>
+  </head>
+  <body>
+    <script>
+      ;(function(root, factory) {
+        console.log('没有模块环境，直接挂载在全局对象上')
+        root.umdModule = factory()
+      })(this, function() {
+        return {
+          bootstrap: async (props) => {
+            console.log('bootstrap2')
+          },
+          mount: async (props) => {
+            console.log('mount2')
+          },
+          unmount: async (props) => {
+            console.log('unmount2')
+          }
+        }
+      })
+    </script>
+  </body>
+</html> -->
